dtngw/requirements.txt
certifi==2023.7.22
influxdb-client==1.38.0
paho-mqtt==1.6.1
python-dateutil==2.8.2
reactivex==4.0.4
six==1.16.0
typing_extensions==4.8.0
urllib3==2.0.7
